qmllint: Properly warn about missing required properties in delegates
Previously we would often suggest to add an id to the view containing
the delegate so you could access the model that way which is very
misleading and might be especially bad as an automatic suggestion.

This change now recommends that the user adds a required property,
although we currently lack the logic to properly insert one for the user.

Fixes: QTBUG-103101
Change-Id: I4cd858ff6617d12b9abf070f524dc626d699e4f1
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>