Allow specifies the native texture format on QQuickRenderTarget
Add some overloaded function. If you need to use VkImage correctly, Must
know the format it corresponds to the VkImage. In the QQuickRenderTarget,
It's use the QRhi::newTexture get a QRhiTexture object, and use the
QRhiTexture::createFrom to create texture use VkImage and VkImageLayout,
on default, it's presume the texture format is QRhiTexture::RGBA8, but
the VkImage is source from user, its format is indeterminate, so we
should allow to specifies the format of VkImage from user.

[ChangeLog][QtQuick] Added an overload for
QQuickRenderTarget::fromOpenGLTexture to allow specifying the format of
the OpenGL texture.

[ChangeLog][QtQuick] Added an overload for
QQuickRenderTarget::fromD3D11Texture to allow specifying the format of
the D3D11 texture.

[ChangeLog][QtQuick] Added an overload for
QQuickRenderTarget::fromMetalTexture to allow specifying the format of
the Metal texture.

[ChangeLog][QtQuick] Added an overload for
QQuickRenderTarget::fromVulkanImage to allow specifying the format of
the Vulkan texture.

Fixes: QTBUG-103216
Change-Id: I3b90be9bbd37439f906351782cf08c060d98cded
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>