QML: Add markers corresponding to various QML attributes
QML attributes are transitive: a base type's attributes leak
into derived type

This causes issues in type registration since we cannot tell
whether a type has a specific attribute is really present
for it. For instance, things like QML_EXTENDED, QML_ATTACHED,
QML_SINGLETON and a few others are the offenders

The way to overcome this, in the lack of the own type, is to
store a static meta object of the enclosing type as a marker
when we encounter a particular attribute declaration. We can
then attest whether the marker stored matches the static meta
object of the type we assess: if so, it is the same type;
if not, we are in a derived type. Consider:

class X : public QObject {
  Q_OBJECT

  QML_EXTENDED
  //-> unpacks roughly in:
  //   (old) using QmlExtendedType = ...;
  //   (new) inline constexpr void marker() {}
};

class Y : public X {
  Q_OBJECT
};

Before:
X::QmlExtendedType - exists
Y::QmlExtendedType - exists (due to inheritance)
After:
X::QmlExtendedType - exists
Y::QmlExtendedType - exists
classType<decltype(X::marker)> == X // truly QmlExtendedType
classType<decltype(Y::marker)> != Y // falsely QmlExtendedType

Note that this commit does not use the markers, but only
introduces them

Change-Id: I22fdee056c2fc63bc64386dd5de62e6021f15a02
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>