Baseline: add more client properties
This is a slightly simplified version compared to the widget base line
test, as we don't need to consider the style.

Most importantly, dark vs light mode is encoded in the appearance ID as
a human-readable string, and the platform and OS versions are added in a
way that is compatible to the widgets tests.

Change-Id: I037d63cee8d54e17fcdcf0a1558492b282191046
Reviewed-by: Jenny Lofthus <jenny.lofthus@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>