FileDialog: fix selectedFile being incorrect due to async view changes
To display a list of files in FileDialog, we use a ListView with a
FolderListModel as the model. Part of being able to respect the
selectedFile that the user has set involves setting that ListView's
currentIndex to the index of selectedFile in the current folder.

When currentFolder was set to a directory that contained a lot of
files, there was an issue where we set the ListView's currentIndex
before it had finished loading its items, resulting in an incorrect
selected item.

To fix this, we now store a "pending" currentIndex, and listen to
the ListView's countChanged signal to know when we can set that
pending currentIndex. This is not an ideal solution, but it's the
most efficient one given the limited information we have from
ListView.

A better solution might be to have some way of telling ListView not
reset currentIndex when its model changes (as long as the currentIndex
that was set is within the bounds of the new model).

Fixes: QTBUG-103547
Change-Id: If3677e5e1a7873a939af19d4ea996ff74033c6e0
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit ae5fad789139489e5497ea4dbf98af4f588108dc)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>