QQmlJS::Dom: replace std::function with qxp::function_ref
There is no std::function_ref, yet, and it's not clear it'll make it
into C++23. At the same time, using std::function is a very coarse
fall-back, as std::function, unlike function_ref, allocates, and
requires the contained function object to be copyable.

P0792r9, unlike std::function, supports const and noexcept modifiers
(<int(int) const noexcept>), and requires a const when a const
function object is passed, so the normal use case is expected to
include the const. This patch only adds it where the compiler would
complain when absent. Since std::function doesn't understand const and
noexcept qualifiers that way, strip them from the signature for the
std::function fall-back.

There seems to be a bug with older MSVC versions (<= 2019). Given that
MSVC 2022 and other compilers are not affected, simply fallback to
std::function. As all APIs are internal, this will not cause potential
ABI issues.

Task-number: QTBUG-103739
Change-Id: Idf1c15280f9e418a093f37fb349396c6703e3c31
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>