QQmlDebug: reliably print the debugger warning
When an executable contains some TUs that pass printWarning and some
that don't, the warning could have been lost if static initialization
first initialized one that had printWarning=false and only then moved
to initializing one that had printWarning=true. Typically, a DLL might
pass false here, in which case the user application, initialized
later, wouldn't have a say in whether the warning was printed.

Use an atomic_flag to independently track whether the warning was
printed, so that we reliably print the warning when at least one TU in
the final executable requested it.

[ChangeLog][QtQml][QQmlDebug] The warning about enabled
debuggers is now printed when at least one translation unit in the
final executable requests it, independent of the order in which
translation units are linked/initialized.

Pick-to: 6.3 6.2 5.15
Change-Id: I10af0a46ecb82a8b1a1373eb9332d913c03b20f3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>