Fix race condition on QQmlEnginePrivate::qml_debugging_enabled
C++11 allows static dynamic initialization from different TUs to
happen concurrently, which means the QQmlDebuggingEnabler ctor must be
re-entrant and synchronized with other users of qml_debugging_enabled.

Thankfully, this is just a flag, so the fix is to simply make it
atomic<> and use relaxed loads and stores on it.

Pick-to: 6.3 6.2 5.15
Change-Id: I0305ab55be86a0e286016a3d1d97ee9bc0e28070
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>