qmlimportscanner: Add logging to help with debugging performance
Add two new logging categories:
 qt.qml.import.scanner
 qt.qml.import.scanner.files

The first one logs how many files are scanned, which files are scanned
timestamps for operations done on each file, durations for the operations
that we expect to take longer.

The second category logs each qml file that is processed recursively
(so not only the input files). Because the output is quite verbose,
it's a separate category rather than being part of the first one.

This is useful to identify performance bottlenecks in projects, without
having to profile the executable.

Pick-to: 6.2 6.3
Task-number: QTBUG-103187
Change-Id: I09506b488dda594b371d229e5fd535a534fe9515
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>