Fix \relates for Q_IMPORT_QML_PLUGIN
Refer to the class instead of trying to refer to the header. The latter
does not associate it with the actual page and creates a proxy page
instead.
As a drive-by, mention that the macro only exists from 6.2 onwards.

Change-Id: I531ee92257e6631a5b7830d881b1ef42a7ea61b1
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
(cherry picked from commit feff2898fe6b457ed9f999c110bb89dea042285b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>