Fix QQuickStackLayout::rearrange for newly added items
Removes redundant calls to ensureLayoutItemsUpdated()

Add childItemsChanged(), which will keep count and currentIndex
properties updated, and emitting the corresponding signals for those
properties whenever children are added or removed.
Notice that these two properties will be updated immediately (not until
we get a updatePolish() call).
Only size hints and rearranges might be scheduled later until a
updatePolish() call.

This also makes the implementation of updateLayoutItems() empty,
basically because there is no shadow structure for holding the list of
layout items in this layout, so to keep that list of layout items
"updated" is therefore not needed.

Change-Id: Ibf8232ea1ad41a3d96b25bd2ea6aab6b5018e3bb
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>