QQuickTableView: fix deprecated [=] capture of *this
Reach *this through d->q_func() instead, so it doesn't need to be captured.

Amends de1672713ae413b563eefec37d1aad155d78a1f8.

An alternative would be to use [&], but the function is long and I was
too lazy to review whether capture-by-reference would have the same
effect as capture-by-value.

Change-Id: If595e9b5c9c64e5a807dc999c3f0c228ff9f2cce
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>