QQuickColorDialog: Implement ColorInputs in C++
The previous version of the ColorInputs.qml contained a good number of
signal handlers that executed js code.
We prefer to write business logic in C++ since it's faster, and we have
more control. Which is the main focus of this patch.

Additionally, it works around an issue where the binding on the
visibility property for the alpha text fields would cause them to be
omitted from the RowLayout's implicitWidth calculation, that would then be
used by the StackLayout to calculate its own implicitWidth in a
bottom-up direction. This made the size of the dialog become larger when
a different selection was chosen in the combobox, and had a negative
impact on the user experience. (The initial selection is the hexcode
text field, which isn't inside a RowLayout, like the other children of
the StackLayout).

The only workaround that I managed to find, was to call setVisibility
form C++, which is why this patch serves a dual purpose.

Some other minor QoL changes are also included. (Like the footer having
a separator for eating left over space).

The most significalt part, user experience wise, is that the ColorInputs
component will now have a "minimum" width, which will impact the size of
the rest of the dialog.
The reason I did it like that, was because I wanted all of the text
fields to be able to fit 4 characters inside of them, without causing
any scrolling/eliding/clipping. The text fields would simply be too
small to fix 4 characters for some of the styles. This means that this
patch affects the size of the dialog for some styles.

Fixes: QTBUG-103900
Pick-to: 6.4
Change-Id: Ic88156c0ab1e692ee38eda29d103c9d2adea7b47
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>