qmltyperegistrar: Extract into own library
Split qmltyperegistrar in two parts: the tool (commandline parsing etc)
and the lib (that does the actual qmltyperegistration generation).
The tool lives at tools/qmltyperegistrar and the lib lives in
src/qmltyperegistrar.
This would allow to test qmltyperegistrar and is a first step into adding
some structure to qmltyperegistrar.

Moves the qqmljsstreamwriter from QmlCompiler to QmlTypeRegistrar.
Also, moves the qmltypes.prf from the lib to the tool, so it can still
be used from qmake projects!

Fixes: QTBUG-103862
Change-Id: I8627d1181db139d043228a802f047a35bff63bb5
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>