CMake: Turn qmlls into an app instead of a tool as a workaround
A developer can have a host Qt without a qmlls, but still want to
cross-compile Qt including all tools (the Yocto SDK scenario).

While configuring the target Qt, they would get

  Searching for tool 'Qt6::qmlls' in package Qt6QmlTools.
  CMake Error at qtbase/cmake/QtToolHelpers.cmake:170 (message):
  The tool "Qt6::qmlls" was not found in the Qt6QmlTools package.
  Package found: 1

qmlls does not contribute to the Qt build itself (like moc or other
code generators), so a host qmlls is not required to build a target
Qt.
But qt_internal_add_tool currently does not support not requiring
the host tool when cross-building it.

Work around this by building qmlls with qt_internal_add_app.
Place it in libexec because it's not user facing and don't
cross-compile it for platforms that don't support the SDK scenario.

Pick-to: 6.3
Task-number: QTBUG-100047
Change-Id: I4678ea791b21ecf3a57ccc870b9a4be09e2a8950
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>