QQuickTreeViewDelegate: toggle expanded on press
When TreeView is interactive, it should expand
tree nodes on mouse release (to not interfere with flicking).
Otherwise, if interactive is false, it should expand already
on mouse press (equal to how e.g Widgets work).

The current if-test that implemented this logic was
broken, and would fail when TreeView was not interactive.

Change-Id: I801a694b2dfd03875893bb2bda5c8ef3883dec75
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>