Add convenience API for making QQmlDataTest-based tests fail on warnings
After this patch, if a QQmlDataTest-derived class passes
FailOnWarningsPolicy::FailOnWarnings to the base constructor, any
non-empty warning encountered by that test will result in a test
failure.

This avoids the need to duplicate the catch-all regex in tests that
want to fail on warnings.

The goal is to gradually enable failure-on-warnings over time.

Leave comments and explicitly pass DoNotFailOnWarnings for tests that
should never fail on warnings.

Task-number: QTBUG-98718
Pick-to: 6.2 6.3 6.4
Change-Id: I4b647d93a0f28ac891c4bdb19ef74569f2918e8f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>