tst_qqmlnotifier: Do not assume qfatal crashes the process
Because we run a test, we might have a handler for SIGABRT installed. If
qfatal uses SIGABRT to terminate the process, we will thus actually
handle it gracefully.
Instead, simply check that the process has terminated, and rely on the
output being available.
As we check for "FATAL" in the output, the exit status check was anyway
redundant before.

Fixes: QTBUG-104051
Change-Id: I062a592e97e717cd1a0d93f7913b492aa322296a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 27cc99809d532e422c0be8c3238e0fde1d6964a8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>