tst_qqmlnotifier: Do not assume qfatal crashes the process
Because we run a test, we might have a handler for SIGABRT installed. If
qfatal uses SIGABRT to terminate the process, we will thus actually
handle it gracefully.
Instead, simply check that the process has terminated, and rely on the
output being available.
As we check for "FATAL" in the output, the exit status check was anyway
redundant before.

Fixes: QTBUG-104051
Pick-to: dev 6.3
Change-Id: I062a592e97e717cd1a0d93f7913b492aa322296a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>