Ensure QmlSingleton attribute does not leak into derived types in C++
Use the QmlTypeHasMarker helper struct to disable Qml type attributes
for types that are not marked with the corresponding macros but merely
inherit them from a superclass.

This fixes the use of the QML_SINGLETON macro, other commits will follow
for the other attributes.

[ChangeLog][QtQml][Important Behavior Changes] Classes decorated with the QML_SINGLETON macro will
stop passing on the QML_SINGLETON functionality to their subclasses.

Task-number: QTBUG-104178
Change-Id: I3adf3e11b734b64ea3c3d2540d1e33963e979a04
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>