Ensure QmlForeign attribute does not leak into derived types in C++
Use the QmlTypeHasMarker helper struct to disable Qml type attributes
for types that are not marked with the corresponding macros but merely
inherit them from a superclass.

This fixes the use of the QML_FOREIGN macro, other commits will follow
for the other attributes.

[ChangeLog][QtQml][Important Behavior Changes] Classes decorated with the QML_FOREIGN macro will
stop passing on the QML_FOREIGN functionality to their subclasses.

Task-number: QTBUG-104179
Change-Id: Ib394e8568bca29aff10561557c1af3574b405d2a
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>