Fix rendernode example wrt stacking
...for OpenGL. This means that putting opaque content on top (which uses
the opaque pass of the sg renderer) is now working as it should.
Previously the custom rendering was not testing against the what was in
the depth buffer, which was fine for geometry drawn in the alpha pass,
but not for others. The associated bug report's example with adding a
Button on top presumably triggers the latter.

The software backend has no problems.

The D3D12 and Metal code paths are no longer supported and will not be
updated anymore. Those anyway feature various limitations, such as a
lack of clipping support, so having limited stacking, depending on the
content on top, is acceptable.

Fixes: QTBUG-102954
Change-Id: I0ee97fb9790ee43f005b1de8226ef760c56c2573
Reviewed-by: Andy Nichols <andy.nichols@qt.io>