Handle unprocessed pressed case in Flickable
Handle case where Flickable filters a pressed event before the receiver
has decided whether it wants to keep the grab. That causes the Flickable
to not process the press event, but still processes all the later
events.
This happens when the receiver is MouseArea so add special case
for it in Flickable, where we start the drag if all conditions are met.

Fixes: QTBUG-38765
Fixes: QTBUG-74842
Change-Id: I2e02042cb496e9eb3b7b5da824e56f76ee76ccbb
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit d7b5a485583004ad6a1374a50b7c3f6cab00aca3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>