Add support for separate blending factors in QSGMaterialShader
This needs to also have an opt-in flag, because existing code must
continue to work as-is (i.e. when it does not touch the new
srcAlpha/dstAlpha fields at all, the color blending factors should be
used for alpha too). Therefore, by default there is no change in
behavior, no matter if one touches the alpha blending factors or not.
The renderer will continue to use the color factors for everything.
Only when the flag is flipped will srcAlpha and dstAlpha have an
effect.

Inspired by a recent text node bug which, while solved differently in
the end, shows that the ability to specify separate factors is useful
in practice occasionally.

[ChangeLog][QtQuick] When a custom material declares its own blending
factors, it is now possible to specify the source/destination alpha
blending factors separate from the color blending factors. This is
done by first setting the separateBlendFactors to true in
QSGMaterialShader::GraphicsPipelineState, and then writing the desired
values to srcAlpha and dstAlpha, in addition to srcColor and dstColor.

Change-Id: I83826a68b80d76500ee195db73b29d393a7d0381
Reviewed-by: Andy Nichols <andy.nichols@qt.io>