Fix alpha with native text rendering
The immediate problem with incorrect semi-transparent Text with
NativeRendering is a porting error from 5.14 times, comparing with

https://code.qt.io/cgit/qt/qtdeclarative.git/tree/src/quick/scenegraph/qsgdefaultglyphnode_p.cpp?h=5.15#n292

shows that the constant color used for blending does not have alpha
premultiplied into the color in 5.15.

In the rhi-based code updateUniformData does the premultiplication when
setting the data for the uniform, like in Qt 5, but then in
upgradeGraphicsPipelineState it also multiplies whereas Qt 5 did not.

In fact that there's a Q_UNUSED(state) at the top of the function
shows that accessing state.opacity() was not meant to be there in the
first place and got added later for reasons unknown.

Fixes: QTBUG-100820
Change-Id: I3acb839676c00e7aa1e2ff2f93eefcaa87b7d0de
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 08e2a90f9afd0a7852f4f594029fa21d2fc1478e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>