Avoid QQuickWindowImpl space leak
The reader for the `screen` property of
`QQuickWindowImpl::screen` always creates a new `QQuickScreenInfo`
instance to wrap the required information about the current screen that
the window sits on.

For example, in the following code:

```
import QtQuick
import QtQuick.Window

Window {
    Component.onCompleted: {
        screen;
        screen;
    }
}
```

The reader function will be called twice and two instances of
`QQuickScreenInfo` will be created and parented to the window.

To avoid creating a new instance for every read of the property and to
avoid cluttering the object tree of the window, ensure that
`QQuickWindowQmlImpl::screen` stops creating a new instance on each
call, instead re-using a previously created instance when available.

To ensure that the `QQuickScreenInfo` instance that is used for the
`screen` property has up-to-date information on the screen, ensure that
the `QQuickScreenInfo` instance that is held by a window is recreated
after a screen change, by deleting a previously created instance on a
`screenChanged` signal.

Add some test-cases to `tst_qquickwindow` to ensure that multiple reads
of the `screen` property do not clutter the object tree and a change of
screen invalidates the `QQuickScreenInfo` instance that the window holds
after a read.

Add an additional test case to ensure that bindings are correctly
updated after a change of screen and the deletion of a screen info
instance.

To correctly perform such a test we need to have two screens available
in the test environment. To ensure that this is the case we use the
`offscreen` plugin, which allows a customization of the available
screens through its configuration.

As we need to create a custom QApplication with a specific platform
plugin and a specific configuration, considering that the current
`tst_qquickwindow` already provides an implicit application to its
tests, a new test suite "tst_qquickwindow_appless" was added to support
the case of constructing the configured application.

Fixes: QTBUG-104404
Fixes: QTBUG-127065
Change-Id: I7e48fe66a7133d3d26f15b097ca77137cfb2728a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>