QQuickAbstractDialog: emit rejected() on mere close()
Like QDialog, there should be only two ways to close a
QQuickAbstractDialog: by accepting or rejecting. Closing a dialog with
the [X] button or hitting Esc should reject. This is natural, and what
QDialog has been implementing since decades, but QQuickAbstractDialog
only emitted rejected() on a manual reject().

To implement, move the emission of accepted() and rejected() to
close() instead of done(), and swap the calls to close() and
setResult() in done() so the correct result() is available when
close() is called.

[ChangeLog][Important Behavior Changes] QtQuick dialogs now emit
rejected() on a mere close(), too.

[ChangeLog][QtQuickDialogs][QQuickAbstractDialog] Now emits rejected()
on a mere close(), too.

The code in 5.15 is too different, so I made a separate fix.

Pick-to: 6.4 6.3 6.2
Fixes: QTBUG-101973
Change-Id: I90d1a78ee95b72eecae35005e191312c939cecf3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>