Fix debugger crash on attach
The debugger treated QQmlGadgetPtrWrapper like normal QObjects,
which means that their metacalls where done incorrectly.
The meta system saw the QQmlGadgetPtrWrapper's metaobject,
which is actually the wrapped type, and produced a garbage
QObject* by reinterpret_casting the wrapper as the wrapped type.
The debugger was changed to recognize gadgets and to call
the read/write/resetOnGadget() methods instead of calling
QMetaProperty.read/write/reset().
Added a test with the qml example that was making the
application crash when a qml debugger was attached to it.

Fixes: QTBUG-104379
Change-Id: I7c1138fc44ee78137386dcd7da145bbaab574afc
Pick-to: 6.3 6.4
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>